// src/app/common/models/nutrition.ts
import { Nutrient } from './nutrient';

export interface Nutrition {
  PROCNT: Nutrient;
  FAT: Nutrient;
  CHOCDF: Nutrient;
  ENERC_KCAL: Nutrient;
  SUGAR: Nutrient;
  FIBTG: Nutrient;
  CA: Nutrient;
  FE: Nutrient;
  P: Nutrient;
  K: Nutrient;
  NA: Nutrient;
  VITA_IU: Nutrient;
  TOCPHA: Nutrient;
  VITD: Nutrient;
  VITC: Nutrient;
  VITB12: Nutrient;
  FOLAC: Nutrient;
  CHOLE: Nutrient;
  FATRN: Nutrient;
  FASAT: Nutrient;
  FAMS: Nutrient;
  FAPU: Nutrient;
}